move temp components
