+++ packages/mars-theme/src/components/components/breadcrumbs/index.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { connect, styled, decode } from 'frontity';
+import { connect, decode } from 'frontity';
 import Link from '../link';
 import Item from './item';
 import Divider from './divider';
-import { isSearchLink } from '../heplers/content';
+import { isSearchLink } from '../../helpers/content';
 import CategoryItem from './category-item';
+
+import { Wrapper } from './styles';
 
 const Breadcrumbs = ({ state }) => {
   const data = state.source.get(state.router.link);
@@ -54,20 +56,8 @@ const Breadcrumbs = ({ state }) => {
   );
 };
 
-export default connect(Breadcrumbs);
-
-const Wrapper = styled.div`
-  width: 100%;
-  font-size: 0.75rem;
-  line-height: 1.3125rem;
-  margin-bottom: 20px;
-  a {
-    &:hover {
-      color: #658022;
-    }
-  }
-`;
-
 Breadcrumbs.propTypes = {
   state: PropTypes.object,
 };
+
+export default connect(Breadcrumbs);
