+++ packages/mars-theme/src/components/components/breadcrumbs/styles.js
@@ -1,0 +1,14 @@
+import { styled } from 'frontity';
+import theme from '../../theme';
+
+export const Wrapper = styled.div`
+  width: 100%;
+  font-size: 0.75rem;
+  line-height: 1.3125rem;
+  margin-bottom: 20px;
+  a {
+    &:hover {
+      color: #658022;
+    }
+  }
+`;
