+++ packages/mars-theme/src/components/components/misc/entity-info/authors-list.js
