+++ packages/mars-theme/src/components/components/misc/entity-info/category-description.js
