+++ packages/mars-theme/src/components/components/misc/entity-info/description.js
