+++ packages/mars-theme/src/components/components/misc/entity-info/expanded-description.js
