+++ packages/mars-theme/src/components/components/misc/entity-info/index.js
